Move MockChainTip to zebra-chain
