+++ zebra-chain/Cargo.toml
@@ -9,7 +9,7 @@ edition = "2018"
 
 [features]
 default = []
-proptest-impl = ["proptest", "proptest-derive", "zebra-test", "rand", "rand_chacha"]
+proptest-impl = ["proptest", "proptest-derive", "zebra-test", "rand", "rand_chacha", "tokio"]
 bench = ["zebra-test"]
 
 [dependencies]
@@ -60,6 +60,7 @@ proptest-derive = { version = "0.3.0", optional = true }
 
 rand = { version = "0.8", optional = true }
 rand_chacha = { version = "0.3", optional = true }
+tokio = { version = "1.15.0", optional = true }
 
 # ZF deps
 ed25519-zebra = "3.0.0"
